refactor: drop reference contributors, implement new goto handler
The previous implementation conflicts with webTypes.
