+++ src/main/resources/META-INF/plugin.xml
@@ -15,11 +15,8 @@
         <polySymbols.webTypes
                 source="META-INF/webTypes.json"
                 enableByDefault="true"/>
-        <xml.tagNameProvider
-                implementation="com.github.tempest.framework.ComponentTagNameProvider"/>
-        <psi.referenceContributor
-                language="HTML"
-                implementation="com.github.tempest.framework.ComponentReferenceContributor"/>
+        <gotoDeclarationHandler
+                implementation="com.github.tempest.framework.ComponentGotoDeclarationHandler"/>
         <spellchecker.bundledDictionaryProvider
                 implementation="com.github.tempest.framework.SpellcheckingDictionaryProvider"/>
     </extensions>
